+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+<a name="3.8.1"></a>
+## [3.8.1](https://github.com/mengxiong10/vue2-datepicker/compare/v3.8.0...v3.8.1) (2020-11-26)
+
+
+### Bug Fixes
+
+* `input` event work with slot input  ([#546](https://github.com/mengxiong10/vue2-datepicker/issues/546)) ([ef1a748](https://github.com/mengxiong10/vue2-datepicker/commit/ef1a748))
+* clear text error ([#544](https://github.com/mengxiong10/vue2-datepicker/issues/544)) ([bcc4f77](https://github.com/mengxiong10/vue2-datepicker/commit/bcc4f77))
+* click event compatible with xhtml ([#550](https://github.com/mengxiong10/vue2-datepicker/issues/550)) ([b7cbc0d](https://github.com/mengxiong10/vue2-datepicker/commit/b7cbc0d))
+* showWeekNumber don't work ([57f2219](https://github.com/mengxiong10/vue2-datepicker/commit/57f2219))
+
+
+
+<a name="3.8.0"></a>
+# [3.8.0](https://github.com/mengxiong10/vue2-datepicker/compare/v3.7.0...v3.8.0) (2020-11-18)
+
+
+### Bug Fixes
+
+* the MMM parse error ([c504b1d](https://github.com/mengxiong10/vue2-datepicker/commit/c504b1d))
+
+
+### Features
+
+* add event `panel-change` ([3d29b49](https://github.com/mengxiong10/vue2-datepicker/commit/3d29b49))
+* add event `update:show-time-panel` ([#531](https://github.com/mengxiong10/vue2-datepicker/issues/531)) ([45969a1](https://github.com/mengxiong10/vue2-datepicker/commit/45969a1))
+* add prop `getYearPanel` ([#526](https://github.com/mengxiong10/vue2-datepicker/issues/526)) ([7fc299a](https://github.com/mengxiong10/vue2-datepicker/commit/7fc299a))
+
+
+
+<a name="3.7.0"></a>
+# [3.7.0](https://github.com/mengxiong10/vue2-datepicker/compare/v3.6.3...v3.7.0) (2020-10-20)
+
+
+### Features
+
+* add prop `formatter` to replace the object usage of `format` ([5d3d547](https://github.com/mengxiong10/vue2-datepicker/commit/5d3d547))
+
+
+
+<a name="3.6.3"></a>
+## [3.6.3](https://github.com/mengxiong10/vue2-datepicker/compare/v3.6.2...v3.6.3) (2020-10-17)
+
+
+
 <a name="3.6.2"></a>
 ## [3.6.2](https://github.com/mengxiong10/vue2-datepicker/compare/v3.6.1...v3.6.2) (2020-08-10)